django-rest/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.11
django-configurations==2.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-nose==1.4.7
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.4.0
idna==3.6
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
nose==1.3.7
openapi-codec==1.3.2
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tblib==3.0.0
tomli==2.0.1
toposort==1.10
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
